requirements/test.txt
-r base.txt
aiosqlite==0.12.0
flake8==3.7.9
isort==4.3.21
pytest-aiohttp==0.3.0
pytest-cov==2.8.1
